refactor: split UI tests into separate files
Split monolithic WasmUiSmokeTests.cs into 4 focused test classes:
- AuthTests: API health + login
- HomePageTests: sync flow
- PageNavigationTests: all page load tests (9 pages)
- TransactionsTests: transaction page + smoke test
